pyproject.toml
[project]
name = "persona-evals"
version = "0.1.0"
description = "Evaluate LLM personas"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "backoff==2.2.1",
    "l2m2==0.0.56",
    "nltk==3.9.1",
    "numpy==2.2.5",
    "pydantic==2.11.4",
    "pymupdf==1.25.5",
    "pymupdf4llm==0.0.17",
    "python-dotenv==1.1.0",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.1.1",
    "ruff>=0.11.6",
]


[tool.mypy]
exclude = ["build", "dist", "tests"]

ignore_missing_imports = true
explicit_package_bases = true
namespace_packages = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
disallow_any_generics = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true

mypy_path = ["."]
python_version = "3.13"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
target-version = "py313"
line-length = 120
lint.select = ["E", "F", "W", "B"]
